refactor: post util functions Move all util functions to a `utils` file to clean up the saga file.
